pyproject.toml
[tool.poetry]
name = "jenezis"
version = "0.1.0"
description = "Knowledge Graph System for Talent Intelligence"
authors = ["Julien Dabert <jdabert@sigilum.fr>"]
readme = "README.md"
license = "Proprietary"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9"
pandas = ">=2.3.2,<3.0.0"
openpyxl = ">=3.1.5,<4.0.0"
openai = ">=1.107.3,<2.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
fastapi = ">=0.116.2,<0.117.0"
uvicorn = ">=0.35.0,<0.36.0"
gunicorn = ">=22.0.0,<23.0.0"
requests = ">=2.32.5,<3.0.0"
prometheus-client = ">=0.23.1,<0.24.0"
alembic = "^1.16.5"
psycopg2-binary = "^2.9.10"
pgvector = "^0.4.1"
sqlalchemy = "^2.0.43"
asyncpg = "^0.30.0"
celery = "^5.5.3"
redis = "^6.4.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyyaml = "^6.0.1"
jinja2 = "^3.1.2"
neo4j = "5.24.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.4.2,<9.0.0"
pytest-cov = ">=5.0.0,<6.0.0"
pytest-asyncio = ">=0.23.0,<0.24.0"
pytest-mock = ">=3.14.0,<4.0.0"
httpx = ">=0.28.1,<0.29.0"
black = ">=24.10.0,<25.0.0"
ruff = ">=0.8.0,<0.9.0"
mypy = ">=1.13.0,<2.0.0"
pre-commit = ">=3.5.0,<4.0.0"
types-requests = "*"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.ruff]
line-length = 100
target-version = "py39"
exclude = [
    ".venv",
    "build",
    "dist",
    "__pycache__",
    "migrations",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise ... from err
    "SIM108", # use ternary operator
    "E722",  # bare except statements (some are intentional)
]

[tool.ruff.lint.isort]
known-first-party = ["src", "tests"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true
ignore_missing_imports = true
exclude = [
    ".venv",
    "build",
    "dist",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/migrations/*",
    "*/config.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
